move pipes in to separate files for stackblitz
