+++ apps/docs/src/app/core/component-docs/table/table-docs.module.ts
@@ -6,10 +6,7 @@ import { API_FILES } from '../../api-files';
 import { TableDocsHeaderComponent } from './table-docs-header/table-docs-header.component';
 import { TableDocsComponent } from './table-docs.component';
 import { TableExampleComponent } from './examples/table-example.component';
-import {
-    FilterTableByPipe, SortTableByPipe,
-    TableColumnSortingExampleComponent
-} from './examples/table-column-sorting-example.component';
+import { TableColumnSortingExampleComponent } from './examples/table-column-sorting-example.component';
 import { TableCdkExampleComponent } from './examples/table-cdk-example.component';
 import { TableResponsiveExampleComponent } from './examples/table-responsive-example.component';
 import { TableCheckboxesExampleComponent } from './examples/table-checkboxes-example.component';
@@ -19,6 +16,8 @@ import { TableFooterExampleComponent } from './examples/table-footer-example.com
 import { TableActivableExampleComponent } from './examples/table-activable-example.component';
 import { TableSemanticExampleComponent } from './examples/table-semantic-example.component';
 import { TablePopinExampleComponent } from './examples/table-popin-example/table-popin-example.component';
+import { SortTableByPipe } from './examples/table-example-sort.pipe';
+import { FilterTableByPipe } from './examples/table-example-filter.pipe';
 
 const routes: Routes = [
     {
@@ -32,8 +31,15 @@ const routes: Routes = [
 ];
 
 @NgModule({
-    imports: [RouterModule.forChild(routes), SharedDocumentationModule, TableModule, CheckboxModule, ObjectStatusModule, ListModule],
-    exports: [RouterModule],
+    imports: [
+        RouterModule.forChild(routes),
+        SharedDocumentationModule,
+        TableModule,
+        CheckboxModule,
+        ObjectStatusModule,
+        ListModule
+    ],
+    exports: [RouterModule, SortTableByPipe, FilterTableByPipe],
     declarations: [
         TableDocsComponent,
         TableExampleComponent,
